Bug 1824856 - import docker image definitions from firefox-android
This switches the android-components docker image to a debian 12 base,
although it could maybe stay on ubuntu 22.04 instead.

Since commit
https://github.com/mozilla/application-services/commit/ac9107b988529ba48e400dc34e781f31d9739626,
application-services builds using an ubuntu 22.04 image, and android-components
consumes the nimbus-fml binary built from that, so running on debian 11 fails
due to an older libc.

Because pip in debian 12 refuses to work outside a virtualenv, pass
`--break-system-packages` to silence the error.

Differential Revision: https://phabricator.services.mozilla.com/D201418